add query-response for get-all-posts

+++ src/features/blogs/application/blogs.service.ts
@@ -29,12 +29,19 @@ export class BlogsService {
     }
 
     async deleteBlog(id: string) {
+        const findedBlog = await this.blogModel.findById(id)
+        if (!findedBlog) {
+            throw new NotFoundException(`Post with id ${id} not found`)
+        }
         const deleteBlog = await this.blogModel.deleteOne({_id: id})
         return deleteBlog
     }
 
     async findBlogById(id: string) {
         const findedBlog = await this.blogModel.findById(id)
+        if (!findedBlog) {
+            throw new NotFoundException(`Post with id ${id} not found`)
+        }
         return findedBlog
     }
